Test changes based on feedback
 * Moved the command factory to common factories
 * Separated out the attrs first to make it easier to read
 * Used dict merging and kwargs instead of a function
 * Used more explicit id values for the fixtures
