+++ tests/common/factories/__init__.py
@@ -8,6 +8,7 @@ from .auth_client import AuthClient, ConfidentialAuthClient
 from .auth_ticket import AuthTicket
 from .authz_code import AuthzCode
 from .base import set_session
+from .bulk_api_command import BulkAPICommand
 from .document import Document, DocumentMeta, DocumentURI
 from .feature import Feature
 from .flag import Flag
@@ -26,6 +27,7 @@ __all__ = (
     "AuthClient",
     "AuthTicket",
     "AuthzCode",
+    "BulkAPICommand",
     "ConfidentialAuthClient",
     "DeveloperToken",
     "Document",
